+++ src/Stratis.Features.FederatedPeg.Tests/SignedMultisigTransactionBroadcasterTests.cs
@@ -63,7 +63,7 @@ namespace Stratis.Features.FederatedPeg.Tests
 
             // Setup MempoolManager.
             this.dateTimeProvider = Substitute.For<IDateTimeProvider>();
-            this.nodeSettings = new NodeSettings(networksSelector: FederatedPegNetwork.NetworksSelector, protocolVersion: NBitcoin.Protocol.ProtocolVersion.ALT_PROTOCOL_VERSION);
+            this.nodeSettings = new NodeSettings(networksSelector: CirrusNetwork.NetworksSelector, protocolVersion: NBitcoin.Protocol.ProtocolVersion.ALT_PROTOCOL_VERSION);
 
             this.mempoolSettings = new MempoolSettings(this.nodeSettings)
             {
