+++ src/Stratis.Features.FederatedPeg.Tests/WithdrawalExtractorTests.cs
@@ -31,7 +31,7 @@ namespace Stratis.Features.FederatedPeg.Tests
 
         public WithdrawalExtractorTests()
         {
-            this.network = FederatedPegNetwork.NetworksSelector.Regtest();
+            this.network = CirrusNetwork.NetworksSelector.Regtest();
             this.counterChainNetwork = Networks.Stratis.Regtest();
 
             this.loggerFactory = Substitute.For<ILoggerFactory>();
