--- src/Stratis.Sidechains.Networks/CirrusV2/CirrusTest.cs
@@ -1,152 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using NBitcoin;
-using NBitcoin.DataEncoders;
-using NBitcoin.Protocol;
-using Stratis.Bitcoin.Features.PoA;
-using Stratis.Bitcoin.Features.SmartContracts.PoA;
-using Stratis.SmartContracts.Networks.Policies;
-
-namespace Stratis.Sidechains.Networks.CirrusV2
-{
-    public class CirrusTest : PoANetwork
-    {
-        /// <summary> The name of the root folder containing the different federated peg blockchains.</summary>
-        private const string NetworkRootFolderName = "cirrus";
-
-        /// <summary> The default name used for the federated peg configuration file. </summary>
-        private const string NetworkDefaultConfigFilename = "cirrus.conf";
-
-        internal CirrusTest()
-        {
-            this.Name = nameof(CirrusTest);
-            this.NetworkType = NetworkType.Testnet;
-            this.CoinTicker = "TCRS";
-            this.Magic = 0x522357B;
-            this.DefaultPort = 26179;
-            this.DefaultMaxOutboundConnections = 16;
-            this.DefaultMaxInboundConnections = 109;
-            this.DefaultRPCPort = 26175;
-            this.DefaultAPIPort = 38223;
-            this.MaxTipAge = 2 * 60 * 60;
-            this.MinTxFee = 10000;
-            this.FallbackFee = 10000;
-            this.MinRelayTxFee = 10000;
-            this.RootFolderName = NetworkRootFolderName;
-            this.DefaultConfigFilename = NetworkDefaultConfigFilename;
-            this.MaxTimeOffsetSeconds = 25 * 60;
-
-            var consensusFactory = new SmartContractPoAConsensusFactory();
-
-            // Create the genesis block.
-            this.GenesisTime = 1544113232;
-            this.GenesisNonce = 56989;
-            this.GenesisBits = new Target(new uint256("0000ffff00000000000000000000000000000000000000000000000000000000"));
-            this.GenesisVersion = 1;
-            this.GenesisReward = Money.Zero;
-
-            string coinbaseText = "https://news.bitcoin.com/markets-update-cryptocurrencies-shed-billions-in-bloody-sell-off/";
-            Block genesisBlock = CirrusNetwork.CreateGenesis(consensusFactory, this.GenesisTime, this.GenesisNonce, this.GenesisBits, this.GenesisVersion, this.GenesisReward, coinbaseText);
-
-            this.Genesis = genesisBlock;
-
-            // Configure federation public keys used to sign blocks.
-            // Keep in mind that order in which keys are added to this list is important
-            // and should be the same for all nodes operating on this network.
-            var genesisFederationMembers = new List<IFederationMember>()
-            {
-                new CollateralFederationMember(new PubKey("03e89abd3c9e791f4fb13ced638457c85beb4aff74d37b3fe031cd888f0f92989e"), new Money(0), null),
-                new CollateralFederationMember(new PubKey("026b7b9092828f3bf9e73995bfa3547c3bcd3814f8101fac626b8349d9a6f0e534"), new Money(0), null),
-                new CollateralFederationMember(new PubKey("02a8a565bf3c675aee4eb8585771c7517e358708faee4f9db2ed7502d7f9dae740"), new Money(0), null),
-                new CollateralFederationMember(new PubKey("0248de019680c6f18e434547c8c9d48965b656b8e5e70c5a5564cfb1270db79a11"), new Money(0), null),
-                new CollateralFederationMember(new PubKey("034bd1a94b0ae315f584ecd22b2ad8fa35056cc70862f33e3e08286f3bbe2207c4"), new Money(0), null),
-            };
-
-            var consensusOptions = new PoAConsensusOptions(
-                maxBlockBaseSize: 1_000_000,
-                maxStandardVersion: 2,
-                maxStandardTxWeight: 100_000,
-                maxBlockSigopsCost: 20_000,
-                maxStandardTxSigopsCost: 20_000 / 5,
-                genesisFederationMembers: genesisFederationMembers,
-                targetSpacingSeconds: 16,
-                votingEnabled: false,
-                autoKickIdleMembers: false
-            );
-
-            var buriedDeployments = new BuriedDeploymentsArray
-            {
-                [BuriedDeployments.BIP34] = 0,
-                [BuriedDeployments.BIP65] = 0,
-                [BuriedDeployments.BIP66] = 0
-            };
-
-            var bip9Deployments = new NoBIP9Deployments();
-
-            this.Consensus = new Consensus(
-                consensusFactory: consensusFactory,
-                consensusOptions: consensusOptions,
-                coinType: 400,
-                hashGenesisBlock: genesisBlock.GetHash(),
-                subsidyHalvingInterval: 210000,
-                majorityEnforceBlockUpgrade: 750,
-                majorityRejectBlockOutdated: 950,
-                majorityWindow: 1000,
-                buriedDeployments: buriedDeployments,
-                bip9Deployments: bip9Deployments,
-                bip34Hash: new uint256("0x000000000000024b89b42a942fe0d9fea3bb44ab7bd1b19115dd6a759c0808b8"),
-                ruleChangeActivationThreshold: 1916, // 95% of 2016
-                minerConfirmationWindow: 2016, // nPowTargetTimespan / nPowTargetSpacing
-                maxReorgLength: 240, // 2 loops of PoA members.
-                defaultAssumeValid: null,
-                maxMoney: Money.Coins(20_000_000),
-                coinbaseMaturity: 1,
-                premineHeight: 2,
-                premineReward: Money.Coins(20_000_000),
-                proofOfWorkReward: Money.Coins(0),
-                powTargetTimespan: TimeSpan.FromDays(14), // two weeks
-                powTargetSpacing: TimeSpan.FromMinutes(1),
-                powAllowMinDifficultyBlocks: false,
-                posNoRetargeting: false,
-                powNoRetargeting: true,
-                powLimit: null,
-                minimumChainWork: null,
-                isProofOfStake: false,
-                lastPowBlock: 0,
-                proofOfStakeLimit: null,
-                proofOfStakeLimitV2: null,
-                proofOfStakeReward: Money.Zero
-            );
-
-            // Same as current smart contracts test networks to keep tests working
-            this.Base58Prefixes = new byte[12][];
-            this.Base58Prefixes[(int)Base58Type.PUBKEY_ADDRESS] = new byte[] { 55 }; // P
-            this.Base58Prefixes[(int)Base58Type.SCRIPT_ADDRESS] = new byte[] { 117 }; // p
-            this.Base58Prefixes[(int)Base58Type.SECRET_KEY] = new byte[] { (239) };
-            this.Base58Prefixes[(int)Base58Type.ENCRYPTED_SECRET_KEY_NO_EC] = new byte[] { 0x01, 0x42 };
-            this.Base58Prefixes[(int)Base58Type.ENCRYPTED_SECRET_KEY_EC] = new byte[] { 0x01, 0x43 };
-            this.Base58Prefixes[(int)Base58Type.EXT_PUBLIC_KEY] = new byte[] { (0x04), (0x35), (0x87), (0xCF) };
-            this.Base58Prefixes[(int)Base58Type.EXT_SECRET_KEY] = new byte[] { (0x04), (0x35), (0x83), (0x94) };
-            this.Base58Prefixes[(int)Base58Type.PASSPHRASE_CODE] = new byte[] { 0x2C, 0xE9, 0xB3, 0xE1, 0xFF, 0x39, 0xE2 };
-            this.Base58Prefixes[(int)Base58Type.CONFIRMATION_CODE] = new byte[] { 0x64, 0x3B, 0xF6, 0xA8, 0x9A };
-            this.Base58Prefixes[(int)Base58Type.STEALTH_ADDRESS] = new byte[] { 0x2b };
-            this.Base58Prefixes[(int)Base58Type.ASSET_ID] = new byte[] { 115 };
-            this.Base58Prefixes[(int)Base58Type.COLORED_ADDRESS] = new byte[] { 0x13 };
-
-            Bech32Encoder encoder = Encoders.Bech32("tb");
-            this.Bech32Encoders = new Bech32Encoder[2];
-            this.Bech32Encoders[(int)Bech32Type.WITNESS_PUBKEY_ADDRESS] = encoder;
-            this.Bech32Encoders[(int)Bech32Type.WITNESS_SCRIPT_ADDRESS] = encoder;
-
-            this.Checkpoints = new Dictionary<int, CheckpointInfo>();
-
-            this.DNSSeeds = new List<DNSSeedData>();
-            this.SeedNodes = new List<NetworkAddress>();
-
-            this.StandardScriptsRegistry = new SmartContractsStandardScriptsRegistry();
-
-            Assert(this.Consensus.HashGenesisBlock == uint256.Parse("0x00008460b940e3e9c7415a07a54cb569a9f69adf790961f11de0c42aa6470708"));
-            Assert(this.Genesis.Header.HashMerkleRoot == uint256.Parse("0xb68311ebfee717754de683570de6e792a2149776381ed49df9cdf3383e59749d"));
-        }
-    }
-}
